Move storage reconciliation to its own file
